pyproject.toml
[tool.poetry]
name = "ha-on-railway"
version = "0.1.0"
description = ""
authors = ["Denis Sakhno <97180503+ideniska@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "ha_on_railway"}]


[tool.poetry.dependencies]
python = "^3.10"
Django = "4.1"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
gunicorn = "^20.1.0"
celery = "^5.2.7"
django-cors-headers = "^3.13.0"
drf-yasg = "^1.21.3"
redis = "^4.3.4"
dj-rest-auth = "^2.2.5"
djangorestframework-simplejwt = "~=5.2"
stripe = "^4.1.0"
python-telegram-bot = "^13.14"
flower = "^1.2.0"
django-celery-beat = "^2.4.0"
PyAMQP = "^0.1.0.7"
Faker = "^15.3.4"
cffi = "^1.15.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
pytest-django = "~=4.5"
pytest-mock = "~=3.8"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit = [
  "*/migrations/*",
  "*/settings*",
  "manage.py",
  "*/apps.py",
]

[tool.coverage.report]
exclude_lines = [
  "^\\s*@(abc.)?abstractmethod",
  "^\\s*@(typing.)?overload",
  "^\\s*if (typing.)?TYPE_CHECKING:",
  "^\\s*if (settings.)?DEBUG:",
  "pragma: no ?cover",
  "def __repr__",
  "def __str__",
  "if self.debug:",
  "raise AssertionError",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
]

[tool.coverage.xml]
output = "coverage.xml"
